+++ GUI-dev/src/models/models.ts
@@ -35,7 +35,6 @@ export class StoreProcessRating {
 }
 
 
-
 export class Review {
   // define props...
   review_id?: number;